# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful backend for a blog client. Handles posts, image uploads, and URL metadata extraction for Editor.js integration.",
        "title": "Blog API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/whitenest/chapters/{number}": {
            "get": {
                "description": "Returns the chapter with the given serial number along with\nminimal references to the previous and next chapters, if any.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "whitenest"
                ],
                "summary": "Get a Whitenest chapter by number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chapter number (1-indexed)",
                        "name": "number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.WhitenestChapterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dtos.CategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dtos.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dtos.ErrorDetail"
                }
            }
        },
...
        "dtos.PostResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "category": {
                    "$ref": "#/definitions/dtos.CategoryResponse"
                },
                "category_id": {
                    "type": "integer"
                },
                "content": {
                    "$ref": "#/definitions/models.EditorJsContent"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.TagResponse"
                    }
                },
                "title": {
                    "type": "string"
                },
                "total_views": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "whitenest_chapter_number": {
                    "type": "integer"
                }
            }
        },
        "dtos.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "dtos.TagResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dtos.WhitenestChapterRef": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "whitenest_chapter_number": {
                    "type": "integer"
                }
            }
        },
        "dtos.WhitenestChapterResponse": {
            "type": "object",
            "properties": {
                "chapter": {
                    "$ref": "#/definitions/dtos.PostResponse"
                },
                "next": {
                    "$ref": "#/definitions/dtos.WhitenestChapterRef"
                },
                "previous": {
                    "$ref": "#/definitions/dtos.WhitenestChapterRef"
                }
            }
        },
        "models.EditorJsBlock": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.EditorJsContent": {
            "type": "object",
            "properties": {
                "blocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EditorJsBlock"
                    }
                },
                "time": {
                    "description": "Unix timestamp in milliseconds",
                    "type": "integer"
                },
                "version": {
                    "description": "Editor.js version",
                    "type": "string"
                }
            }
        },
...
    }
}